Made ierr output-only.
In the fortran function, ierr does double-duty, serving as an input
flag (1 means intput was factored, 0 means it was not), and as an
output flag. In this code, the function is always being supplied
with unfactored input (ierr was always initially set to 0) so I
revised the pp_def to call the fortran function with an input of
zero, and then to set ierr to the revised value.
ierr was doing double-duty